requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
